hotfix(Tag): 매칭 생성 - 매칭별 태그 저장 로직 수정

See merge request s12-fintech-finance-sub1/S12P21C208!226

+++ backend/src/main/java/com/patriot/fourlipsclover/tag/repository/TagRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TagRepository extends JpaRepository<Tag, Long> {
+public interface TagRepository extends JpaRepository<Tag, Integer> {
 
-	Tag findByName(String tag);
+    Tag findByName(String tag);
 }
